[frontend] fix Artifacts table width
